+++ packages/model/jest.config.ts
@@ -1,4 +1,8 @@
-module.exports = {
+import { pathsToModuleNameMapper } from 'ts-jest';
+/* eslint-disable */
+const { compilerOptions } = require('../../tsconfig.base.json');
+
+export default {
   displayName: 'model',
   preset: '../../jest.preset.js',
   globals: {
@@ -13,4 +17,7 @@ module.exports = {
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
   coverageDirectory: '../../coverage/packages/model',
   testEnvironment: 'node',
+  moduleNameMapper: pathsToModuleNameMapper(compilerOptions.paths, {
+    prefix: '<rootDir>/../..',
+  }),
 };